evals/requirements.txt
azure-ai-evaluation
azure-identity
azure-core
pandas
python-dotenv
matplotlib
seaborn
scikit-learn
semantic-kernel
openai
pydantic
requests
urllib3